+++ packages/rsc/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,136 @@
 
 ### Patch Changes
 
+- Updated dependencies [59fcf30]
+  - ai@6.0.78
+
+## 0.0.1
+
+### Patch Changes
+
+- ai@6.0.77
+
+## 0.0.1
+
+### Patch Changes
+
+- ai@6.0.76
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [7168375]
+  - ai@6.0.75
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [471009b]
+  - ai@6.0.74
+
+## 0.0.1
+
+### Patch Changes
+
+- ai@6.0.73
+
+## 0.0.1
+
+### Patch Changes
+
+- ai@6.0.72
+
+## 0.0.1
+
+### Patch Changes
+
+- ai@6.0.71
+
+## 0.0.1
+
+### Patch Changes
+
+- ai@6.0.70
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [d659305]
+  - ai@6.0.69
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [8bf2660]
+  - ai@6.0.68
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [53f6731]
+  - ai@6.0.67
+
+## 0.0.1
+
+### Patch Changes
+
+- ai@6.0.66
+
+## 0.0.1
+
+### Patch Changes
+
+- ai@6.0.65
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [ce9daa3]
+  - ai@6.0.64
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [be95579]
+  - ai@6.0.63
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [2810850]
+  - ai@6.0.62
+
+## 0.0.1
+
+### Patch Changes
+
+- ai@6.0.61
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [5fc42fa]
+  - ai@6.0.60
+
+## 0.0.1
+
+### Patch Changes
+
+- ai@6.0.59
+
+## 0.0.1
+
+### Patch Changes
+
 - ai@6.0.58
 
 ## 0.0.1